+++ CHANGES.md
@@ -1,5 +1,9 @@
 Poodinis Changelog
 ==================
+** Next version**
+* DEPRECATE ADD_CONCRETE_TYPE_REGISTRATION registration option. It basically does nothing anymore. See next point.
+* CHANGE adding registrations by super type always registers them by concrete type as well now. (Previously done with ADD_CONCRETE_TYPE_REGISTRATION). See DO_NOT_ADD_CONCRETE_TYPE_REGISTRATION for the reverse behaviour.
+
 **Version 4.0.0**
 * REMOVE deprecated module "dependency.d"